Add no-std support for the signature verification cache
- std retains std::sync::RwLock use. no-std uses spin::RwLock. spin is now an unconditional dependency.
- Move and reuse the Lock type used by the revocation status cache
- Remove the feature gate on the Arc blanket impl, which needs only alloc.
- Register the cache in the pittv3 environments that build paths
