+++ pittv3-lib/src/options_std_app.rs
@@ -107,6 +107,10 @@ pub fn options_std_app(args: &Pittv3Args) {
 
     let mut pe = PkiEnvironment::default();
     pe.populate_5280_pki_environment();
+    // The graph and the paths built over it re-present the same CA signatures many times;
+    // caching them is sound because a hit means this exact signature already verified under
+    // this exact key.
+    pe.add_signature_cache(Box::new(DefaultSignatureVerificationCache::new()));
     pe.add_trust_anchor_source(Box::new(ta_store.clone()));
     pe.add_certificate_source(Box::new(cert_source.clone()));
